Started chat, client GUI stub

+++ lib/ts_map/include/ts_map.h
@@ -8,7 +8,7 @@
 
 typedef struct ts_map_node {
     struct ts_map_node *left, *right, *parent;
-    int key;
+    uint64_t key;
     void *val;
 } ts_map_node_t;
 
@@ -21,8 +21,8 @@ typedef struct ts_map {
 void ts_map_init(ts_map_t *m);
 void ts_map_destroy(ts_map_t *m, void (* destructor)(void *));
 
-bool ts_map_insert(ts_map_t *m, int key, void *val);
-void *ts_map_find(ts_map_t *m, int key);
-void *ts_map_erase(ts_map_t *m, int key);
+bool ts_map_insert(ts_map_t *m, uint64_t key, void *val);
+void *ts_map_find(ts_map_t *m, uint64_t key);
+void *ts_map_erase(ts_map_t *m, uint64_t key);
 
 #endif /* XENON_TS_MAP_TS_MAP_H */
